requirements.txt
visdom==0.1.8.9
tqdm==4.62.3
pytorch-msssim==0.2.1
lpips==0.1.4
kornia==0.6.5
dominate==2.6.0
thop==0.1.1.post2209072238
opencv-python==4.5.5.64
pillow==8.2.0
imageio==2.9.0
imageio-ffmpeg==0.5.1
